avoid type cast

+++ opencv_paddle_ocr_test/db_post_process.cc
@@ -71,8 +71,8 @@ cv::RotatedRect Unclip(std::vector<std::vector<float>> const &box,
     offset.Execute(soln, distance);
     std::vector<cv::Point2f> points;
 
-    for (int j = 0; j < soln.size(); j++) {
-        for (int i = 0; i < soln[soln.size() - 1].size(); i++) {
+    for(size_t j = 0; j < soln.size(); j++) {
+        for (size_t i = 0; i < soln[soln.size() - 1].size(); i++) {
             points.emplace_back(soln[j][i].X, soln[j][i].Y);
         }
     }
